refactor: EventCategoryService interface 제거
